requirements.txt
attrs==20.3.0
aws-cdk.assets==1.115.0
aws-cdk.aws-applicationautoscaling==1.115.0
aws-cdk.aws-autoscaling-common==1.115.0
aws-cdk.aws-cloudformation==1.115.0
aws-cdk.aws-cloudwatch==1.115.0
aws-cdk.aws-codeguruprofiler==1.115.0
aws-cdk.aws-codestarnotifications==1.115.0
aws-cdk.aws-ec2==1.115.0
aws-cdk.aws-ecr==1.115.0
aws-cdk.aws-ecr-assets==1.115.0
aws-cdk.aws-efs==1.115.0
aws-cdk.aws-events==1.115.0
aws-cdk.aws-iam==1.115.0
aws-cdk.aws-kms==1.115.0
aws-cdk.aws-lambda==1.115.0
aws-cdk.aws-logs==1.115.0
aws-cdk.aws-s3==1.115.0
aws-cdk.aws-s3-assets==1.115.0
aws-cdk.aws-sagemaker==1.115.0
aws-cdk.aws-signer==1.115.0
aws-cdk.aws-sns==1.115.0
aws-cdk.aws-sqs==1.115.0
aws-cdk.aws-ssm==1.115.0
aws-cdk.cloud-assembly-schema==1.115.0
aws-cdk.core==1.115.0
aws-cdk.custom-resources==1.115.0
aws-cdk.cx-api==1.115.0
aws-cdk.region-info==1.115.0
boto3==1.18.6
botocore==1.21.6
cattrs==1.6.0
-e git+git@ssh.gitlab.aws.dev:vgodwin/comprend_email_custom_classification_project.git@825468908a033dd5de56d2128ba3a5b692794e4f#egg=comprehend_custom_classification
constructs==3.3.101
jmespath==0.10.0
jsii==1.32.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
